🎨 change java module folder & code style
